Update docs, fix -v, make -l optional

+++ repeat.sh
@@ -5,14 +5,13 @@
 set -eu
 [ "${DEBUG:-0}" = "1" ] && set -x
 
-logdir="logs"
-zero="$0"
 sleep_lock_wait="1"
 sleep_after_unlock="1"
 background=0
+out_file='/dev/stdout'
 read_delim=$'\n'
 read_file=
-out_file='/dev/stdout'
+logdir=
 
 
 _exittrap () {
@@ -21,31 +20,40 @@ _exittrap () {
 _info () { echo "$(date) $0 $BASHPID: $*" ; }
 _process_lock () {
     local lock_content="$BASHPID $0 $*"
-	local chksum="$(printf "%s\n" "$lock_content" | md5sum - | awk '{print $1}')"
-    local result
+    local result chksum
 
-	while [ -e "$logdir/$chksum.lock" ] ; do
-        read -a lock_data < "$logdir/$chksum.lock"
-		ppid="${lock_data[0]}"
-		cmdname="${lock_data[1]}"
-		if ! kill -0 "$ppid" 2>/dev/null || [ ! "$cmdname" = "$0" ] ; then
-			_info "ppid '$ppid' gone or cmdname '$cmdname' != '$0'; lock stale, removing '$logdir/$chksum.lock'"
-			rm -f "$logdir/$chksum.lock"
-			continue
-		fi
+    if [ -n "$logdir" ] ; then
+        chksum="$(printf "%s\n" "$lock_content" | md5sum - | awk '{print $1}')"
+        while [ -e "$logdir/$chksum.lock" ] ; do
+            read -a lock_data < "$logdir/$chksum.lock"
+            ppid="${lock_data[0]}"
+            cmdname="${lock_data[1]}"
+            if ! kill -0 "$ppid" 2>/dev/null || [ ! "$cmdname" = "$0" ] ; then
+                _info "ppid '$ppid' gone or cmdname '$cmdname' != '$0'; lock stale, removing '$logdir/$chksum.lock'"
+                rm -f "$logdir/$chksum.lock"
+                continue
+            fi
+            _info "Waiting for '$logdir/$chksum.lock'"
+            sleep "$sleep_lock_wait"
+        done
+        echo "$lock_content" > "$logdir/$chksum.lock"
+    fi
 
-		_info "Waiting for '$logdir/$chksum.lock'"
-        sleep "$sleep_lock_wait"
-	done
-
-	echo "$lock_content" > "$logdir/$chksum.lock"
     _info "running command: $*"
-    _info "running command: $*" >> "$logdir/$chksum.log" 2>&1
-	"$@" >>"$logdir/$chksum.log" 2>&1
-    result=$?
-	rm -f "$logdir/$chksum.lock"
+    if [ -n "$logdir" ] ; then
+        _info "running command: $*" >> "$logdir/$chksum.log" 2>&1
+        "$@" >>"$logdir/$chksum.log" 2>&1
+        result=$?
+        echo "" >>"$logdir/$chksum.log"
+        rm -f "$logdir/$chksum.lock"
+        _info "command '$*' exited with status $result" >> "$logdir/$chksum.log" 2>&1
+    else
+        "$@" 2>&1
+        result=$?
+        echo ""
+    fi
     _info "command '$*' exited with status $result"
-    _info "command '$*' exited with status $result" >> "$logdir/$chksum.log" 2>&1
+
     sleep "$sleep_after_unlock"
 }
 _process_loop () {
@@ -78,10 +86,11 @@ Usage: $0 [OPTIONS] [COMMAND [ARGS ..]]
 
 Continuously runs COMMAND (and any ARGS) in a loop.
 
-A lock file is used to track what commands are still running, and STDOUT and
-STDERR are redirected to a log file. The lock file contains the PID, process
-name, and command and arguments being executed. The log and lock file names are
-the MD5 checksum of the contents of the lock file.
+If '-l' is passed, a lock file is used to track what commands are still running,
+and STDOUT and STDERR are redirected to a log file. The lock file contains the
+PID, process name, and command and arguments being executed. The log and lock
+file names are the MD5 checksum of the contents of the lock file. The same lock
+and log file names are used for the duration of a repeat.sh run (per command).
 
 If '-f' is passed, reads a file and passes each line as a final argument to
 COMMAND. If '-f' is passed but no COMMAND is passed, executes each line of the
@@ -92,7 +101,7 @@ Options:
   -W SEC        Seconds to sleep in between checking for a lock to be released. ($sleep_lock_wait)
   -A SEC        Seconds to sleep after a command completes and the lock is removed. ($sleep_after_unlock)
   -f FILE       Read commands from FILE. If '-', reads from standard input.
-  -l DIR        The directory to keep log and lock files. ($logdir)
+  -l DIR        Output logs and lock files to DIR.
   -o FILE       Output the loop's warnings/errors to a FILE. ($out_file)
   -0            If a FILE was passed, separate each entry by the null byte, not newlines
   -b            Fork the command(s) into the background and exit
@@ -116,7 +125,7 @@ while getopts "W:A:f:l:o:0bhv" args ; do
         0)  read_delim=$'\0' ;;
         b)  background=1 ;;
         h)  _usage ;;
-        v)  export DEBUG=1 ;;
+        v)  DEBUG=1 ; set -x ;;
         *)
             echo "$0: Error: unknown option $args" ;
             exit 1 ;;
@@ -128,6 +137,8 @@ if [ -z "$read_file" ] && [ $# -lt 1 ] ; then
     _usage
 fi
 
-[ -d "$logdir" ] || mkdir -p "$logdir"
+if [ -n "$logdir" ] && [ ! -d "$logdir" ] ; then
+    mkdir -p "$logdir"
+fi
 
 _process "$@"
